api/requirements.txt
fastapi>=0.115.12
openai
pydantic>=2.11.4
uvicorn>=0.34.2
python-multipart>=0.0.18
mangum>=0.18.0
python-dotenv

